requirements.txt
aliyun-python-sdk-core==2.13.24
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-debug-toolbar==2.2
django-redis==4.12.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.1
Pillow==7.2.0
PyMySQL==0.10.0
pytz==2020.1
qiniu==7.2.8
redis==3.5.3
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
